+++ src/console_c.c
@@ -1473,8 +1473,8 @@ static void TCOD_console_read_apf(TCOD_console_t con,FILE *f,int width, int heig
 }
 
 static int string_ends_with(const char *str, const char *suffix) {
-	int str_len = strlen(str);
-	int suffix_len = strlen(suffix);
+	size_t str_len = strlen(str);
+	size_t suffix_len = strlen(suffix);
 	return
 		(str_len >= suffix_len) &&
 		(0 == strcmp(str + (str_len-suffix_len), suffix));
